+++ ouzel/audio/WavePlayer.hpp
@@ -19,7 +19,7 @@ namespace ouzel
             {
             }
 
-            void getSamples(uint32_t frames, uint16_t channels, uint32_t sampleRate, std::vector<float>& samples) override
+            void getSamples(uint32_t frames, uint32_t channels, uint32_t sampleRate, std::vector<float>& samples) override
             {
             }
