+++ ouzel/audio/coreaudio/CAAudioDevice.cpp
@@ -87,9 +87,9 @@ namespace ouzel
 
             AudioDevice::AudioDevice(uint32_t initBufferSize,
                                      uint32_t initSampleRate,
-                                     uint16_t initChannels,
+                                     uint32_t initChannels,
                                      const std::function<void(uint32_t frames,
-                                                              uint16_t channels,
+                                                              uint32_t channels,
                                                               uint32_t sampleRate,
                                                               std::vector<float>& samples)>& initDataGetter):
                 audio::AudioDevice(Driver::CoreAudio, initBufferSize, initSampleRate, initChannels, initDataGetter)
@@ -116,7 +116,7 @@ namespace ouzel
                 }
 
                 if (channels > maxChannelCount)
-                    channels = static_cast<uint16_t>(maxChannelCount);
+                    channels = static_cast<uint32_t>(maxChannelCount);
 #elif TARGET_OS_MAC
                 constexpr AudioObjectPropertyAddress deviceListAddress = {
                     kAudioHardwarePropertyDevices,
