+++ ouzel/audio/mixer/Object.hpp
@@ -79,7 +79,7 @@ namespace ouzel
                 }
 
                 // TODO: make non-virtual
-                virtual void getSamples(uint32_t frames, uint16_t channels, uint32_t sampleRate, std::vector<float>& samples)
+                virtual void getSamples(uint32_t frames, uint32_t channels, uint32_t sampleRate, std::vector<float>& samples)
                 {
                     for (auto child : children)
                         child->getSamples(frames, channels, sampleRate, samples);
