+++ ouzel/audio/openal/OALAudioDevice.cpp
@@ -84,9 +84,9 @@ namespace ouzel
 
             AudioDevice::AudioDevice(uint32_t initBufferSize,
                                      uint32_t initSampleRate,
-                                     uint16_t initChannels,
+                                     uint32_t initChannels,
                                      const std::function<void(uint32_t frames,
-                                                              uint16_t channels,
+                                                              uint32_t channels,
                                                               uint32_t sampleRate,
                                                               std::vector<float>& samples)>& initDataGetter):
                 audio::AudioDevice(Driver::OpenAL, initBufferSize, initSampleRate, initChannels, initDataGetter)
@@ -111,7 +111,7 @@ namespace ouzel
                 }
 
                 if (channels > maxChannelCount)
-                    channels = static_cast<uint16_t>(maxChannelCount);
+                    channels = static_cast<uint32_t>(maxChannelCount);
 #endif
 
                 const ALCchar* deviceName = alcGetString(nullptr, ALC_DEFAULT_DEVICE_SPECIFIER);
